Split geolocation file into modules.
- Split modules out to make it easier to test.
- Cleaned up minor code smells.
